requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
scikit-image>=0.19.0
joblib>=1.1.0

# Computer vision and image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Web interface
gradio>=3.0.0

# Progress bars and utilities
tqdm>=4.62.0

# Jupyter notebook support (optional, for training)
jupyter>=1.0.0
ipykernel>=6.0.0

# Data visualization (optional, for analysis)
matplotlib>=3.5.0
seaborn>=0.11.0

# Additional utilities
pathlib2>=2.3.0; python_version < "3.4"